Add hugo-installer to site package.json
This installs a local copy of the Hugo binary that we need to build
the site. Otherwise the user would have to install it via their sy-
stem package manager. This also adds a new npm "build:site" script
to build the site using our local version of Hugo.

Note: this is slightly annoying, but after toying with "easier" op-
tions like Hexo that integrate with the JavaScript ecosystem more
naturally, I ended up preferring Hugo anyways. Hexo's not nearly as
powerful or well documented as Hugo.

Note note: there is also the `hugo-bin` package on NPM, but it is
very out of date.