pyproject.toml: update dependencies
Most of these are minor and would have been selected by the semant-
ic version string during `poetry install`, but I want to make sure
that they are as current as possible before I leave the project.

In the case of Pandas 1.4.0 the minimum Python version is actually
only 3.8, so let's set that as the minimum.