site/layouts: use bubble buttons for elements
I got rid of these when initially trying to group the elements for
each module together under their clusters, but actually now there
are less elements so we can keep this bubble style. A bonus is that
we don't need to figure out how to delimit by comma in Hugo's Go
templates (the last one is always tricky).